Restructure classes for better organization
Restructure classes to create LLMHandler and RAGHandler, and update app.py accordingly.

* **Add LLMHandler class**
  - Create `LLMHandler` class in `backend/llm_handler.py` to store an instance of the Fireworks model.

* **Add RAGHandler class**
  - Create `RAGHandler` class in `backend/rag_handler.py` to have instances of `LLMHandler` and `EmbeddingStorage`.
  - Move the `answer_query` method from `EmbeddingStorage` to `RAGHandler`.

* **Update EmbeddingStorage class**
  - Remove the `answer_query` method from `EmbeddingStorage` in `backend/embedding_storage.py`.
  - Update imports to remove `ChatFireworks` and related dependencies.

* **Update app.py**
  - Import `RAGHandler` and initialize it with `LLMHandler` and `EmbeddingStorage` instances.
  - Update the `/search` endpoint to use the `answer_query` method from `RAGHandler`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/anupmanekar/jira-rag-app/tree/main?shareId=XXXX-XXXX-XXXX-XXXX).
